combiner/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
dateparser==1.1.8
Django==4.2.3
django-crispy-forms==2.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycryptodome==3.18.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
